Add Bob's Buddy Overlay (#52)
This commit introduces the Bob's Buddy overlay to the Twitch Extension.
It allows viewers to see simulation results during Battlegrounds, as
provided by Hearthstone Deck Trackers. Streamers may choose to enable
this feature, and to allow or disallow their viewers to view the results
during certain phases of the game.

Co-authored-by: Alex Zeier <alex@hearthsim.net>